# Conflicts:
#	src/main/java/at/fhburgenland/entities/Besucher.java
#	src/main/java/at/fhburgenland/services/FuehrungService.java
#	src/main/java/at/fhburgenland/services/FuetterungsplanService.java
#	src/main/java/at/fhburgenland/services/GehegeService.java
#	src/main/java/at/fhburgenland/services/GesundheitsakteService.java
#	src/main/java/at/fhburgenland/services/InventarService.java
#	src/main/java/at/fhburgenland/services/NahrungsartService.java
#	src/main/java/at/fhburgenland/services/PflegerService.java
#	src/main/java/at/fhburgenland/services/TierService.java

+++ src/main/java/at/fhburgenland/services/FuehrungService.java
@@ -1,6 +1,6 @@
 package at.fhburgenland.services;
 
-import at.fhburgenland.entities.Fuehrung;
+import at.fhburgenland.entities.*;
 import jakarta.persistence.*;
 import jakarta.validation.ConstraintViolationException;
 
@@ -14,7 +14,18 @@ public class FuehrungService {
         EntityManager em = emf.createEntityManager();
         EntityTransaction et = null;
         try {
-            et = em.getTransaction(); et.begin();
+            et = em.getTransaction();
+            et.begin();
+
+            for (Integer pid : pflegerIds) {
+                Pfleger p = em.getReference(Pfleger.class, pid);
+                f.addPfleger(p);
+            }
+            for (Integer bid : besucherIds) {
+                Besucher b = em.getReference(Besucher.class, bid);
+                f.addBesucher(b);
+            }
+
             em.persist(f);
             et.commit();
         } catch (Exception e){
@@ -63,7 +74,11 @@ public class FuehrungService {
         try{
             et = em.getTransaction(); et.begin();
             Fuehrung f = em.find(Fuehrung.class, id);
-            em.remove(f);
+            if (f != null) {
+                f.getVeranstalter().clear();
+                f.getBesucherListe().clear();
+                em.remove(f);
+            }
             et.commit();
         } catch (Exception e){
             if(et != null) et.rollback();
